refactor dto reserved identifier predicates
