+++ crates/rulemorph/src/dto/support/identifiers.rs
@@ -4,6 +4,10 @@ use crate::dto::DtoLanguage;
 
 use super::{capitalize, lower_camel, pascal_case, snake_case, words_from_key};
 
+mod reserved;
+
+use self::reserved::is_reserved;
+
 pub(in crate::dto) fn field_identifier(
     lang: DtoLanguage,
     key: &str,
@@ -77,327 +81,4 @@ pub(in crate::dto) fn safe_type_name(lang: DtoLanguage, name: &str) -> String {
         ident.push_str("Record");
     }
     ident
-}
-
-fn is_reserved(lang: DtoLanguage, ident: &str) -> bool {
-    match lang {
-        DtoLanguage::Rust => is_reserved_rust(ident),
-        DtoLanguage::TypeScript => is_reserved_typescript(ident),
-        DtoLanguage::Python => is_reserved_python(ident),
-        DtoLanguage::Go => is_reserved_go(ident),
-        DtoLanguage::Java => is_reserved_java(ident),
-        DtoLanguage::Kotlin => is_reserved_kotlin(ident),
-        DtoLanguage::Swift => is_reserved_swift(ident),
-    }
-}
-
-fn is_reserved_rust(value: &str) -> bool {
-    matches!(
-        value,
-        "as" | "break"
-            | "const"
-            | "continue"
-            | "crate"
-            | "else"
-            | "enum"
-            | "extern"
-            | "false"
-            | "fn"
-            | "for"
-            | "if"
-            | "impl"
-            | "in"
-            | "let"
-            | "loop"
-            | "match"
-            | "mod"
-            | "move"
-            | "mut"
-            | "pub"
-            | "ref"
-            | "return"
-            | "self"
-            | "Self"
-            | "static"
-            | "struct"
-            | "super"
-            | "trait"
-            | "true"
-            | "type"
-            | "unsafe"
-            | "use"
-            | "where"
-            | "while"
-    )
-}
-
-fn is_reserved_typescript(value: &str) -> bool {
-    matches!(
-        value,
-        "break"
-            | "case"
-            | "catch"
-            | "class"
-            | "const"
-            | "continue"
-            | "debugger"
-            | "default"
-            | "delete"
-            | "do"
-            | "else"
-            | "enum"
-            | "export"
-            | "extends"
-            | "false"
-            | "finally"
-            | "for"
-            | "function"
-            | "if"
-            | "import"
-            | "in"
-            | "instanceof"
-            | "new"
-            | "null"
-            | "return"
-            | "super"
-            | "switch"
-            | "this"
-            | "throw"
-            | "true"
-            | "try"
-            | "typeof"
-            | "var"
-            | "void"
-            | "while"
-            | "with"
-            | "as"
-            | "implements"
-            | "interface"
-            | "let"
-            | "package"
-            | "private"
-            | "protected"
-            | "public"
-            | "static"
-            | "yield"
-            | "any"
-            | "boolean"
-            | "number"
-            | "string"
-            | "symbol"
-            | "type"
-            | "from"
-            | "of"
-    )
-}
-
-fn is_reserved_python(value: &str) -> bool {
-    matches!(
-        value,
-        "False"
-            | "None"
-            | "True"
-            | "and"
-            | "as"
-            | "assert"
-            | "async"
-            | "await"
-            | "break"
-            | "class"
-            | "continue"
-            | "def"
-            | "del"
-            | "elif"
-            | "else"
-            | "except"
-            | "finally"
-            | "for"
-            | "from"
-            | "global"
-            | "if"
-            | "import"
-            | "in"
-            | "is"
-            | "lambda"
-            | "nonlocal"
-            | "not"
-            | "or"
-            | "pass"
-            | "raise"
-            | "return"
-            | "try"
-            | "while"
-            | "with"
-            | "yield"
-    )
-}
-
-fn is_reserved_go(value: &str) -> bool {
-    matches!(
-        value,
-        "break"
-            | "default"
-            | "func"
-            | "interface"
-            | "select"
-            | "case"
-            | "defer"
-            | "go"
-            | "map"
-            | "struct"
-            | "chan"
-            | "else"
-            | "goto"
-            | "package"
-            | "switch"
-            | "const"
-            | "fallthrough"
-            | "if"
-            | "range"
-            | "type"
-            | "continue"
-            | "for"
-            | "import"
-            | "return"
-            | "var"
-    )
-}
-
-fn is_reserved_java(value: &str) -> bool {
-    matches!(
-        value,
-        "abstract"
-            | "assert"
-            | "boolean"
-            | "break"
-            | "byte"
-            | "case"
-            | "catch"
-            | "char"
-            | "class"
-            | "const"
-            | "continue"
-            | "default"
-            | "do"
-            | "double"
-            | "else"
-            | "enum"
-            | "extends"
-            | "final"
-            | "finally"
-            | "float"
-            | "for"
-            | "goto"
-            | "if"
-            | "implements"
-            | "import"
-            | "instanceof"
-            | "int"
-            | "interface"
-            | "long"
-            | "native"
-            | "new"
-            | "package"
-            | "private"
-            | "protected"
-            | "public"
-            | "return"
-            | "short"
-            | "static"
-            | "strictfp"
-            | "super"
-            | "switch"
-            | "synchronized"
-            | "this"
-            | "throw"
-            | "throws"
-            | "transient"
-            | "try"
-            | "void"
-            | "volatile"
-            | "while"
-    )
-}
-
-fn is_reserved_kotlin(value: &str) -> bool {
-    matches!(
-        value,
-        "as" | "break"
-            | "class"
-            | "continue"
-            | "do"
-            | "else"
-            | "false"
-            | "for"
-            | "fun"
-            | "if"
-            | "in"
-            | "interface"
-            | "is"
-            | "null"
-            | "object"
-            | "package"
-            | "return"
-            | "super"
-            | "this"
-            | "throw"
-            | "true"
-            | "try"
-            | "typealias"
-            | "val"
-            | "var"
-            | "when"
-            | "while"
-    )
-}
-
-fn is_reserved_swift(value: &str) -> bool {
-    matches!(
-        value,
-        "class"
-            | "deinit"
-            | "enum"
-            | "extension"
-            | "func"
-            | "import"
-            | "init"
-            | "let"
-            | "protocol"
-            | "static"
-            | "struct"
-            | "subscript"
-            | "typealias"
-            | "var"
-            | "break"
-            | "case"
-            | "continue"
-            | "default"
-            | "defer"
-            | "do"
-            | "else"
-            | "fallthrough"
-            | "for"
-            | "guard"
-            | "if"
-            | "in"
-            | "repeat"
-            | "return"
-            | "switch"
-            | "where"
-            | "while"
-            | "as"
-            | "Any"
-            | "catch"
-            | "false"
-            | "is"
-            | "nil"
-            | "rethrows"
-            | "super"
-            | "self"
-            | "Self"
-            | "throw"
-            | "throws"
-            | "true"
-            | "try"
-    )
 }
